Decompose pipeline.py: move functions to proper modules
- severity_gte, severity_max → constants.py
- recluster_singletons → clustering.py
- verify_scorer → utils.py
- extract_assessment_rationale → extraction.py
- extract_failing_traces: use search_assessments API
- Simplify _is_non_issue to severity-based check (remove _NO_ISSUE_PATTERNS)
- Update imports and mock paths in tests
